app/requirements.txt
astroid
autopep8
backoff
bcrypt
certifi
cffi
chardet
Click
Flask
Flask-Login
Flask-Misaka
Flask-Session
Flask-Simple-GeoIP
Flask-SQLAlchemy
Flask-WTF
idna
isort
itsdangerous
Jinja2
lazy-object-proxy
mccabe
misaka
pycodestyle
pycparser
pylint
python-dotenv
redis
requests
simple-geoip
six
SQLAlchemy
typed-ast
urllib3
Werkzeug
wrapt
WTForms
email_validator
cryptography
pycryptodome